Add single marker to map
