+++ src/ui/App.css
@@ -29,7 +29,7 @@
   display: flex;
   flex-grow: 1;
   height: 250px;
-  min-width: 320px;
+  min-width: 300px;
   margin: 10px;
 }
